+++ pyprep/utils.py
@@ -64,3 +64,25 @@ def filter_design(N_order, amp, freq):
     )
     kernel = np.multiply(kernel[0 : N_order + 1], (np.transpose(hamming_window[:])))
     return kernel
+
+
+def split_list(mylist, chunk_size):
+    """Split list in chunks.
+
+    Parameters
+    ----------
+    my_list: list
+        list to split.
+    chunk_size: int
+        size of the lists returned.
+
+    Returns
+    -------
+    chunked lists: list
+        list of the chunked lists.
+
+    See: https://stackoverflow.com/a/312466/5201771
+    """
+    return [
+        mylist[offs : offs + chunk_size] for offs in range(0, len(mylist), chunk_size)
+    ]